# docker/proxy/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MDsrv API",
    "version": "1.0.0",
    "description": "Unified public API for modular session, trajectory, topology, and YAML services."
  },
...
  "paths": {
...
    "/api/v1/topology": {
      "get": {
        "tags": ["Topology"],
        "summary": "List stored PDB topology entries",
        "responses": {
          "200": {
            "description": "Topology entries",
            "content": { "application/json": {} }
          }
        }
      },
      "post": {
        "tags": ["Topology"],
        "summary": "Fetch and store a PDB topology from a URL",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url", "id", "fileName"],
                "properties": {
                  "url": { "type": "string" },
                  "id": { "type": "string" },
                  "fileName": { "type": "string" },
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "source": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Invalid request" },
          "409": { "description": "Id or file already exists" }
        }
      }
    },
    "/api/v1/topology/{id}": {
      "put": {
        "tags": ["Topology"],
        "summary": "Upload a PDB topology directly",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "fileName",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "chemical/x-pdb": { "schema": { "type": "string" } },
            "text/plain": { "schema": { "type": "string" } }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Invalid request" },
          "409": { "description": "Id or file already exists" },
          "413": { "description": "Upload exceeds configured limit" }
        }
      },
      "get": {
        "tags": ["Topology"],
        "summary": "Read a stored PDB topology file",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "PDB content",
            "content": { "chemical/x-pdb": {} }
          }
        }
      },
      "patch": {
        "tags": ["Topology"],
        "summary": "Update topology metadata and optionally rename its id/file",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "Optional new id. If provided, the .pdb file is renamed."
                  },
                  "name": { "type": "string", "description": "Display name." },
                  "description": { "type": "string" },
                  "source": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": { "application/json": {} }
          },
          "400": { "description": "Invalid request" },
          "404": { "description": "Missing topology" },
          "409": { "description": "Target id or file exists" }
        }
      },
      "delete": {
        "tags": ["Topology"],
        "summary": "Delete a stored topology",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted" },
          "404": { "description": "Missing topology" }
        }
      }
    },
...
  }
}